+++ src/services/index.ts
@@ -1,12 +1,13 @@
 import * as operatingSystem from './operatingSystem';
 import ForceService from './forceService';
 import configuration from './configuration';
+export { defaultOptions, saveConfigFile } from './configuration';
 
 export { ForceService, operatingSystem, configuration };
 import DXService from './dxService';
 export const dxService = DXService.getInstance();
-import { SObjectCategory } from './dxService';
-export { SObjectCategory };
+import { SObjectCategory, QueryResult, ExecuteAnonymousResult, SFDX } from './dxService';
+export { SObjectCategory, QueryResult, ExecuteAnonymousResult, SFDX };
 import { FCOauth, FCConnection } from './fcConnection';
 import { FCConnectionService } from './fcConnectionService';
 export const fcConnection = FCConnectionService.getInstance();
@@ -22,4 +23,6 @@ import { SaveService } from './saveService';
 export const saveService = SaveService.getInstance();
 import { SaveHistoryService } from './saveHistoryService';
 export const saveHistoryService = SaveHistoryService.getInstance();
+import apexTestResults from './apexTestResults';
+export { apexTestResults };
 export * from './fcZip';
