+++ src/util/index.ts
@@ -1,2 +1,2 @@
-import { isEmptyUndOrNull, saveToFile, removeFile } from './util';
-export { isEmptyUndOrNull, saveToFile, removeFile };
+import { isEmptyUndOrNull, saveToFile, removeFile, toArray } from './util';
+export { isEmptyUndOrNull, saveToFile, removeFile, toArray };
